+++ src/postgres/general.rs
@@ -2,11 +2,13 @@
 
 mod alias;
 mod column;
+mod column_value;
 mod expression;
 mod with;
 
 pub use alias::{Alias, Aliasable};
 pub use column::Column;
+pub use column_value::ColumnValuePair;
 pub use expression::Expression;
 pub use with::{with, WithClause, WithQuery, WithQueryBuilder};